Split crate into log and log_core
