+++ core/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "log_core"
+version = "0.3.4"
+authors = ["The Rust Project Developers"]
+license = "MIT/Apache-2.0"
+repository = "https://github.com/rust-lang/log"
+documentation = "https://doc.rust-lang.org/log/log_core"
+homepage = "https://github.com/rust-lang/log"
+description = """
+A lightweight logging facade for Rust, no_std version.
+"""
+
+[features]
+max_level_off   = []
+max_level_error = []
+max_level_warn  = []
+max_level_info  = []
+max_level_debug = []
+max_level_trace = []
+
+release_max_level_off   = []
+release_max_level_error = []
+release_max_level_warn  = []
+release_max_level_info  = []
+release_max_level_debug = []
+release_max_level_trace = []
+
+no_std = []
